Reorganizing the maven code, adding a parent project and adding readmes
