+++ README.md
@@ -9,7 +9,13 @@ This repository is part of [Apache OpenWhisk](http://openwhisk.incubator.apache.
 
 * [docker-compose](docker-compose/README.md) allows testing OpenWhisk locally, using Docker Compose. This is ideal if you are contributing to core development
 * [node-local](node-local/README.md) allows testing individual OpenWhisk functions locally, using only node.js. This is ideal if you are writing node.js functions to run in OpenWhisk, but need to emulate some of OpenWhisk's behavior in creating `params` and expecting promises.
-* [maven-java](maven-java/README.md) allows testing OpenWhisk Java Actions. This shows how to package the function dependencies e.g. external jar.
+* *java-maven* tooling to support building / deploying Java actions in OpenWhisk
+  * [action-archetype](java-maven/action-archetype) Archetype for creating a Java Action project using Apache Maven
+  * [annotations](java-maven/annotations) Java annotations to indicate OpenWhisk Actions, Packages, Rules and Triggers 
+  * [parent](java-maven/parent) provides a baseline Maven configuration for Maven-based OpenWhisk sub-projects
+  * [update-maven-plugin](java-maven/update-maven-plugin) Apache Maven plugin to read the [annotations](java-maven/annotations) and automatically deploying the Actions, Packages, Rules and Trigger to OpenWhisk
+* [maven-java-examplea](maven-java-example/README.md) shows how to package the function dependencies e.g. external jar.
+* [java-local](java-local) allows testing OpenWhisk Java Actions
 
 ## Travis builds
