+++ java-maven/action-archetype/src/main/resources/META-INF/maven/archetype.xml
