+++ java-maven/action-archetype/src/main/resources/archetype-resources/src/test/java/FunctionAppTest.java
