+++ java-maven/action-archetype/src/test/resources/projects/basic/archetype.properties
