+++ java-maven/action-archetype/src/test/resources/projects/basic/goal.txt
