+++ java-maven/parent/README.md
@@ -1,0 +1,5 @@
+# Apache OpenWhisk Parent
+
+This module is part of the [Apache OpenWhisk](https://openwhisk.incubator.apache.org) project.
+
+This is the parent project for Apache OpenWhisk.
