+++ java-maven/update-maven-plugin/src/main/java/org/apache/openwhisk/maven/commands/TriggerCommand.java
