+++ maven-java-example/dependency-reduced-pom.xml
