--- openwhisk-annotations/pom.xml
@@ -1,23 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.apache</groupId>
-		<artifactId>apache</artifactId>
-		<version>19</version>
-        <relativePath />
-	</parent>
-	
-	<groupId>org.apache.openwhisk</groupId>
-	<artifactId>openwhisk-annotations</artifactId>
-	<version>0.0.1-SNAPSHOT</version>
-	<name>Apache OpenWhisk Annotations</name>
-
-	<properties>
-		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-		<maven.compiler.source>1.8</maven.compiler.source>
-		<maven.compiler.target>1.8</maven.compiler.target>
-	</properties>
-</project>
