--- openwhisk-install-maven-plugin/src/it/simple-it/verify.groovy
@@ -1,3 +1,0 @@
-File touchFile = new File( basedir, "target/touch.txt" );
-
-assert touchFile.isFile()
